requirements.txt
python-telegram-bot>=20.0
yandex-music>=3.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
